Restore test for document filtering
